LayoutTests/inspector/dom/getOuterHTML-expected.txt

Test for DOM.getOuterHTML (Copy HTML).


== Running test suite: DOM.getOuterHTML
-- Running test case: GetOuterHTML
Dumping outerHTML for element div#x:
<div id="x" style="display:none">
    <h1>A Title</h1>
    <p>A Paragraph</p>
</div>
